test(live): move LLM live tests onto Groq/Llama (ADR-0036)
ADR-0036 moved the deployment LLM off gemini-2.5-flash and onto Llama 3.3
70B on Groq, a non-reasoning model with no thinking tail to cap. Align the
live tests with that default:

- Replace the Gemini reasoning_effort A/B (thinking_live_test.go, an
  ADR-0035 artifact superseded for the default path) with a Groq/Llama
  live TTFT-distribution test (latency_live_test.go). Keeping the trivial
  and reasoning-bait tiers turns the old A/B into the live check that
  ADR-0036's "no tail" premise holds: the bait TTFT should track the
  trivial TTFT instead of blowing up the way it did on Gemini. Gated on
  GROQ_API_KEY behind the `live` tag.
- Fix bench-live.yml, which still injected GEMINI_API_KEY while
  TestBench_LiveSLO already required groq.APIKeyEnv (GROQ_API_KEY) — the
  nightly would have failed the key check. Now passes GROQ_API_KEY.
